Restore Encrypted.Binary behaviour
